+++ CHANGELOG.md
@@ -15,3 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Generate Ruby client
 - Add Gemfile
 - Reduce operation id to 70 chars
+
+### Changed
+- Changed api template file
+- Update OpenAPI spec generator tool